[af-316] rename: 경로 변경으로 인한 수정
