+++ src/main/java/com/drunkenlion/alcoholfriday/domain/admin/restaurant/cart/dto/request/RestaurantOrderCartSaveRequest.java
@@ -12,12 +12,12 @@ import java.util.List;
 @AllArgsConstructor(access = AccessLevel.PROTECTED)
 @Schema(description = "레스토랑 발주 장바구니에 제품 추가 요청")
 public class RestaurantOrderCartSaveRequest {
-    @Schema(description = "주문 요청 수량")
-    private Long quantity;
+    @Schema(description = "매장 고유 ID")
+    private Long restaurantId;
 
     @Schema(description = "제품 고유 ID")
     private Long productId;
 
-    @Schema(description = "매장 고유 ID")
-    private Long restaurantId;
+    @Schema(description = "주문 요청 수량")
+    private Long quantity;
 }
