+++ src/main/java/com/drunkenlion/alcoholfriday/domain/admin/restaurant/cart/dto/response/RestaurantOrderCartSaveResponse.java
@@ -1,11 +1,14 @@
 package com.drunkenlion.alcoholfriday.domain.admin.restaurant.cart.dto.response;
 
-import com.drunkenlion.alcoholfriday.domain.restaurant.cart.entity.RestaurantOrderCartDetail;
+import com.drunkenlion.alcoholfriday.domain.admin.restaurant.cart.entity.RestaurantOrderCartDetail;
 import com.drunkenlion.alcoholfriday.global.ncp.dto.NcpFileResponse;
 import io.swagger.v3.oas.annotations.media.Schema;
-import lombok.*;
-
 import java.math.BigDecimal;
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
 
 @Getter
 @Builder
